third_party/blink/web_tests/external/wpt/editing/run/multitest_8001-9000-expected.txt

This is a testharness.js-based test.
Found 47 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["subscript",""],["fontsize","2"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>bar" but got "foo<span style=\\"vertical-align:sub; font-size:small\\">a</span>bar"
[FAIL] [["subscript",""],["fontsize","2"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["fontsize","2"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontsize","2"],["subscript",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>bar" but got "foo<span style=\\"font-size:small; vertical-align:sub\\">a</span>bar"
[FAIL] [["fontsize","2"],["subscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontsize","2"],["subscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["fontsize","3"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"3\\"><sub>a</sub></font>bar" but got "foo<span style=\\"vertical-align:sub\\">a</span>bar"
[FAIL] [["subscript",""],["fontsize","3"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["fontsize","3"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontsize","3"],["subscript",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"3\\"><sub>a</sub></font>bar" but got "foo<span style=\\"vertical-align:sub\\">a</span>bar"
[FAIL] [["fontsize","3"],["subscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontsize","3"],["subscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["hilitecolor","aqua"],["backcolor","tan"],["inserttext","a"]] "foo[]bar" queryCommandValue("hilitecolor") before
  assert_equals: Wrong result returned (after color normalization) expected "rgba(0, 0, 0, 0)" but got "rgb(0, 0, 0)"
[FAIL] [["hilitecolor","aqua"],["backcolor","tan"],["inserttext","a"]] "foo[]bar" queryCommandValue("hilitecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(210, 180, 140)" but got "rgb(0, 0, 0)"
[FAIL] [["backcolor","tan"],["hilitecolor","aqua"],["inserttext","a"]] "foo[]bar" queryCommandValue("hilitecolor") before
  assert_equals: Wrong result returned (after color normalization) expected "rgba(0, 0, 0, 0)" but got "rgb(0, 0, 0)"
[FAIL] [["backcolor","tan"],["hilitecolor","aqua"],["inserttext","a"]] "foo[]bar" queryCommandValue("hilitecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["delete",""],["inserttext","a"]] "foo<b>[bar]</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>a</b>baz" but got "foo<span style=\\"font-weight:700\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<i>[bar]</i>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<i>a</i>baz" but got "foo<span style=\\"font-style:italic\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<s>[bar]</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s>a</s>baz" but got "foo<span style=\\"text-decoration-line:line-through\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sub>[bar]</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>baz" but got "foo<span style=\\"font-size:13.3333px\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sup>[bar]</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>a</sup>baz" but got "foo<span style=\\"font-size:13.3333px\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<u>[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>a</u>baz" but got "foo<span style=\\"text-decoration-line:underline\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com>[bar]</a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>baz" but got "fooabaz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font face=sans-serif>[bar]</font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font face=\\"sans-serif\\">a</font>baz" but got "foo<span style=\\"font-family:sans-serif\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font size=4>[bar]</font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"4\\">a</font>baz" but got "foo<span style=\\"font-size:large\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font color=#0000FF>[bar]</font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">a</font>baz" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><font color=blue>[bar]</font></a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>baz" but got "fooabaz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font color=blue><a href=http://www.google.com>[bar]</a></font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>baz" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><font color=brown>[bar]</font></a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\"><font color=\\"#a52a2a\\">a</font></a>baz" but got "foo<span style=\\"color:rgb(165, 42, 42)\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font color=brown><a href=http://www.google.com>[bar]</a></font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>baz" but got "fooabaz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><font color=black>[bar]</font></a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\"><font color=\\"#000000\\">a</font></a>baz" but got "fooabaz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><u>[bar]</u></a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>baz" but got "fooabaz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<u><a href=http://www.google.com>[bar]</a></u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>baz" but got "foo<span style=\\"text-decoration-line:underline\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sub><font size=2>[bar]</font></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>baz" but got "foo<span style=\\"font-size:small\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font size=2><sub>[bar]</sub></font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"1\\"><sub>a</sub></font>baz" but got "foo<span style=\\"font-size:10.8333px\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sub><font size=3>[bar]</font></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"3\\"><sub>a</sub></font>baz" but got "fooabaz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font size=3><sub>[bar]</sub></font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>baz" but got "foo<span style=\\"font-size:13.3333px\\">a</span>baz"
[FAIL] [["delete",""],["inserttext","a"]] "foo<b>[bar</b>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>a</b>" but got "foo<span style=\\"font-weight:700\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<i>[bar</i>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<i>a</i>" but got "foo<span style=\\"font-style:italic\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<s>[bar</s>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s>a</s>" but got "foo<span style=\\"text-decoration-line:line-through\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sub>[bar</sub>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>" but got "foo<span style=\\"font-size:13.3333px\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sup>[bar</sup>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>a</sup>" but got "foo<span style=\\"font-size:13.3333px\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<u>[bar</u>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>a</u>" but got "foo<span style=\\"text-decoration-line:underline\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com>[bar</a>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>" but got "fooa"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font face=sans-serif>[bar</font>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font face=\\"sans-serif\\">a</font>" but got "foo<span style=\\"font-family:sans-serif\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font size=4>[bar</font>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"4\\">a</font>" but got "foo<span style=\\"font-size:large\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font color=#0000FF>[bar</font>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">a</font>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">a</span>"
Harness: the test ran to completion.
